backend/requirements.txt

fastapi>=0.95.0
uvicorn>=0.21.1
sqlalchemy>=2.0.0
pydantic>=2.0.0
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.6
alembic>=1.10.0
pymysql>=1.0.2
cryptography>=40.0.0
python-dotenv>=1.0.0
